+++ filament/backend/src/opengl/OpenGLProgram.cpp
@@ -111,7 +111,7 @@ OpenGLProgram::OpenGLProgram(OpenGLDriver* gl, const Program& programBuilder) no
         if (programBuilder.hasSamplers()) {
             // if we have samplers, we need to do a bit of extra work
             // activate this program so we can set all its samplers once and for all (glUniform1i)
-            gl->useProgram(program);
+            gl->getContext().useProgram(program);
 
             auto const& samplerGroupInfo = programBuilder.getSamplerGroupInfo();
             auto& indicesRun = mIndicesRuns;
@@ -213,7 +213,7 @@ void OpenGLProgram::updateSamplers(OpenGLDriver* gl) noexcept {
 
             // FIXME: getSampler() is expensive because it's a hashmap lookup
             GLuint sampler = gl->getSampler(samplers[index].s);
-            gl->bindSampler(tmu, sampler);
+            gl->getContext().bindSampler(tmu, sampler);
         }
     }
     CHECK_GL_ERROR(utils::slog.e)
